Remove duplicate `props_helper` module
Utilize the more full-fledged `ember-runtime/tests/props_helper` module
instead.
