+++ packages/ember-metal/tests/accessors/get_test.js
@@ -1,4 +1,4 @@
-import testBoth from 'ember-metal/tests/props_helper';
+import { testBoth } from 'ember-metal/tests/props_helper';
 import {
   get,
   getWithDefault
