+++ packages/ember-metal/tests/watching/unwatch_test.js
@@ -1,4 +1,4 @@
-import testBoth from 'ember-metal/tests/props_helper';
+import { testBoth } from 'ember-metal/tests/props_helper';
 import {
   watch,
   unwatch
